Move message search match
